working specular lighting

+++ OpenGLCourseApp/Light.h
@@ -10,8 +10,8 @@ public:
 	Light(GLfloat red, GLfloat green, GLfloat blue, GLfloat aIntensity, // could also pass in vec with values. These are bound to a shader uniform 
 		GLfloat xDir, GLfloat yDir, GLfloat zDir, GLfloat dIntensity); // these handle diffuse lighting
 
-	void UseLight(GLfloat ambientIntensityLocation, GLfloat ambientColorLocation,
-		GLfloat diffuseIntensityLocation, GLfloat directionLocation);
+	void UseLight(GLuint ambientIntensityLocation, GLuint ambientColorLocation,
+		GLuint diffuseIntensityLocation, GLuint directionLocation);
 
 	~Light();
